Version 2.0
Basic improvements and bug removals: 1) It distinguish between paragraph separating commands (that do not have to be formatted with the text) and paragraph beginning commands (that will be formatted together with the paragraph they begin). 2) Works also on commented text, for any level of comments. The functions are now much more readable, too.