Version 2.3
In this version the comments are formatted even when the comment character is not at the beginning of a line, that is, indented multi-line comments get properly formatted as well.

The comment string is parametrized according to the filetype, so that e.g. formatting works also for c/c++ comments (very few languages supported). If the filetype is not recognized, the comment is set as the tex comment %.

The script also contains a simple but useful Comment command that toggles the comments on a range (again, for few languages only).

Remove or personalize the maps to FormatLatexPar and ToggleComment as desired.